Optimize lock event
