+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/global/database/event/DatabaseAckLockReleasedEvent.java
@@ -32,7 +32,7 @@ public final class DatabaseAckLockReleasedEvent implements GovernanceEvent {
     private final String lockedInstance;
     
     public DatabaseAckLockReleasedEvent(final String ackLockedName) {
-        String[] databaseInstance = LockNodeUtil.parseDatabaseLockName(ackLockedName);
+        String[] databaseInstance = LockNodeUtil.parseAckLockName(ackLockedName);
         this.database = databaseInstance[0];
         this.lockedInstance = databaseInstance[1];
     }
