+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/global/general/ShardingSphereGeneralLockManager.java
@@ -20,16 +20,16 @@ package org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.g
 import com.google.common.eventbus.Subscribe;
 import org.apache.shardingsphere.infra.eventbus.ShardingSphereEventBus;
 import org.apache.shardingsphere.infra.instance.ComputeNodeInstance;
-import org.apache.shardingsphere.infra.lock.LockType;
-import org.apache.shardingsphere.infra.lock.ShardingSphereGlobalLock;
+import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.ShardingSphereGlobalLock;
 import org.apache.shardingsphere.infra.lock.ShardingSphereLock;
-import org.apache.shardingsphere.mode.manager.ShardingSphereLockManager;
+import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.ShardingSphereLockManager;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.LockNodeService;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.LockNodeServiceFactory;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.event.GeneralAckLockReleasedEvent;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.event.GeneralAckLockedEvent;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.event.GeneralLockReleasedEvent;
 import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.event.GeneralLockedEvent;
+import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.util.LockType;
 import org.apache.shardingsphere.mode.persist.PersistRepository;
 import org.apache.shardingsphere.mode.repository.cluster.ClusterPersistRepository;
 
@@ -114,7 +114,7 @@ public final class ShardingSphereGeneralLockManager implements ShardingSphereLoc
      */
     @Subscribe
     public synchronized void locked(final GeneralLockedEvent event) {
-        String lockName = event.getLockName();
+        String lockName = event.getLockedName();
         ShardingSphereGeneralLock lock = locks.get(lockName);
         if (null == lock) {
             lock = createGeneralLock();
@@ -130,7 +130,7 @@ public final class ShardingSphereGeneralLockManager implements ShardingSphereLoc
      */
     @Subscribe
     public synchronized void lockReleased(final GeneralLockReleasedEvent event) {
-        String lockName = event.getLockName();
+        String lockName = event.getLockedName();
         getOptionalLock(lockName).ifPresent(optional -> optional.releaseAckLock(lockName, getCurrentInstanceId()));
     }
     
@@ -141,7 +141,7 @@ public final class ShardingSphereGeneralLockManager implements ShardingSphereLoc
      */
     @Subscribe
     public synchronized void ackLocked(final GeneralAckLockedEvent event) {
-        getOptionalLock(event.getLockName()).ifPresent(optional -> optional.addLockedInstance(event.getLockedInstance()));
+        getOptionalLock(event.getAckLockedName()).ifPresent(optional -> optional.addLockedInstance(event.getLockedInstance()));
     }
     
     /**
@@ -151,7 +151,7 @@ public final class ShardingSphereGeneralLockManager implements ShardingSphereLoc
      */
     @Subscribe
     public synchronized void ackLockReleased(final GeneralAckLockReleasedEvent event) {
-        getOptionalLock(event.getLockName()).ifPresent(optional -> optional.removeLockedInstance(event.getLockedInstance()));
+        getOptionalLock(event.getAckLockedName()).ifPresent(optional -> optional.removeLockedInstance(event.getLockedInstance()));
     }
     
     private String getCurrentInstanceId() {
