+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/global/general/event/GeneralLockReleasedEvent.java
@@ -28,5 +28,5 @@ import org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.Gover
 @Getter
 public final class GeneralLockReleasedEvent implements GovernanceEvent {
     
-    private final String lockName;
+    private final String lockedName;
 }
