+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-standalone-mode/shardingsphere-standalone-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/standalone/lock/ShardingSphereNonReentrantLock.java
@@ -24,7 +24,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Lock;
 
 /**
- * Non reentrant lock implemented ShardingSphereLock.
+ * Non-reentrant lock implemented ShardingSphereLock.
  */
 @RequiredArgsConstructor
 public final class ShardingSphereNonReentrantLock implements ShardingSphereLock {
